Refactor NavbarLogo component and update import paths
- Created a new NavbarLogo component in navbar-logo.tsx, enhancing the logo display with a beta label and theme support.
- Updated import path in navbar.tsx to reflect the new file structure.
- Removed the old NavbarLogo component file to streamline the codebase.
